+++ nufhe/polynomial_transform_ntt.py
@@ -17,8 +17,10 @@
 
 import numpy
 
+from reikna.core import Transformation, Parameter, Annotation, Type
+
 from .transform import ntt1024, ntt1024_requirements, Transform
-from .transform.arithmetic import add, mul, get_ff_elem
+from .transform.arithmetic import add, mul, get_ff_elem, prepare_for_mul, mul_prepared
 from .transform.ntt import ntt_transform_ref
 from .transform import ntt_cpu
 from .performance import PerformanceParametersForDevice
@@ -60,6 +62,13 @@ def transformed_space_mul_ref(data1, data2):
     return ntt_cpu.gnum_to_u64(data1 * data2)
 
 
+def transformed_space_mul_prepared_ref(data1, data2):
+    coeff = ntt_cpu.gnum(0xfffffffe00000001) # Inverse of 2**64 modulo (2**64-2**32+1)
+    data1 = ntt_cpu.gnum(data1)
+    data2 = ntt_cpu.gnum(data2)
+    return ntt_cpu.gnum_to_u64(data1 * data2 * coeff)
+
+
 ff_elem = get_ff_elem()
 
 
@@ -71,8 +80,29 @@ def transformed_mul(perf_params: PerformanceParametersForDevice):
     return mul(ff_elem=ff_elem, method=perf_params.ntt_mul_method).module
 
 
+def transformed_mul_prepared(perf_params: PerformanceParametersForDevice):
+    return mul_prepared(ff_elem=ff_elem, method=perf_params.ntt_mul_method).module
+
+
 def transform_module_requirements():
     return ntt1024_requirements()
+
+
+def get_prepare_for_mul_trf(shape):
+    dtype = transformed_dtype()
+    return Transformation([
+        Parameter('output', Annotation(Type(dtype, shape), 'o')),
+        Parameter('input', Annotation(Type(dtype, shape), 'i'))],
+        """
+        ${dtypes.ctype(dtype)} x = ${input.load_same};
+        ${ff_ctype} x_ff = { x };
+        ${output.store_same}(${prepare_for_mul}(x_ff).val);
+        """,
+        connectors=['input', 'output'],
+        render_kwds=dict(
+            prepare_for_mul=prepare_for_mul(ff_elem=ff_elem).module,
+            dtype=dtype,
+            ff_ctype=transformed_internal_ctype()))
 
 
 def transform_module(perf_params: PerformanceParametersForDevice, multi_iter=False):